+++ pennylane_lightning/lightning_tensor/lightning_tensor.py
@@ -111,7 +111,7 @@ class LightningTensor(Device):
             raise ValueError(f"Unsupported method: {method}")
 
         if shots is not None:
-            raise ValueError("LightningTensor does not support finite shots.")
+            raise ValueError("default.tensor does not support finite shots.")
 
         super().__init__(wires=wires, shots=shots)
 
@@ -143,13 +143,13 @@ class LightningTensor(Device):
         for arg in kwargs:
             if arg not in self._device_options:
                 raise TypeError(
-                    f"Unexpected argument: {arg} during initialization of the LightningTensor device."
+                    f"Unexpected argument: {arg} during initialization of the default.tensor device."
                 )
 
     @property
     def name(self):
         """The name of the device."""
-        return "lightning.tensor"
+        return "default.tensor"
 
     @property
     def num_wires(self):
@@ -266,7 +266,7 @@ class LightningTensor(Device):
             Tuple: The Jacobian for each trainable parameter.
         """
         raise NotImplementedError(
-            "The computation of derivatives has yet to be implemented for the lightning.tensor device."
+            "The computation of derivatives has yet to be implemented for the default.tensor device."
         )
 
     def execute_and_compute_derivatives(
@@ -284,7 +284,7 @@ class LightningTensor(Device):
             tuple: A numeric result of the computation and the gradient.
         """
         raise NotImplementedError(
-            "The computation of derivatives has yet to be implemented for the lightning.tensor device."
+            "The computation of derivatives has yet to be implemented for the default.tensor device."
         )
 
     # pylint: disable=unused-argument
@@ -323,7 +323,7 @@ class LightningTensor(Device):
             tensor-like: A numeric result of computing the vector-Jacobian product.
         """
         raise NotImplementedError(
-            "The computation of vector-Jacobian product has yet to be implemented for the lightning.tensor device."
+            "The computation of vector-Jacobian product has yet to be implemented for the default.tensor device."
         )
 
     def execute_and_compute_vjp(
@@ -344,5 +344,5 @@ class LightningTensor(Device):
             Tuple, Tuple: the result of executing the scripts and the numeric result of computing the vector-Jacobian product
         """
         raise NotImplementedError(
-            "The computation of vector-Jacobian product has yet to be implemented for the lightning.tensor device."
+            "The computation of vector-Jacobian product has yet to be implemented for the default.tensor device."
         )